EntryExit widget and folder restructuring
